requirements.txt
pandas>=2.2.0
numpy>=1.26.4,<2.0.0
pyyaml>=6.0.1
scikit-learn>=1.4.1
fastapi>=0.109.2
uvicorn>=0.27.1
gradio>=4.19.2
notebook>=7.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
xgboost>=2.0.3
mlflow>=2.10.0
optuna>=3.5.0
boto3>=1.34.0
awscli>=1.32.0
dvc>=3.40.0 
dvc-s3>=3.0.0
pydantic>=2.6.0
pytest==8.1.1
httpx==0.27.0


